Create single amazon clients provider
