+++ src/com/sungardas/snapdirector/aws/AmazonConfigProvider.java
@@ -1,0 +1,51 @@
+package com.sungardas.snapdirector.aws;
+
+import com.amazonaws.auth.AWSCredentials;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
+import com.amazonaws.services.ec2.AmazonEC2;
+import com.amazonaws.services.ec2.AmazonEC2Client;
+import com.amazonaws.services.sqs.AmazonSQS;
+import com.amazonaws.services.sqs.AmazonSQSClient;
+
+import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRepositories;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableDynamoDBRepositories(basePackages = "com.sungardas.snapdirector.aws.dynamodb.repository")
+public class AmazonConfigProvider {
+
+    @Value("${amazon.aws.accesskey}")
+    private String amazonAWSAccessKey;
+
+    @Value("${amazon.aws.secretkey}")
+    private String amazonAWSSecretKey;
+
+
+    @Bean
+    public AmazonDynamoDB amazonDynamoDB() {
+        AmazonDynamoDB amazonDynamoDB = new AmazonDynamoDBClient(amazonAWSCredentials());
+        return amazonDynamoDB;
+    }
+
+    @Bean
+    public AWSCredentials amazonAWSCredentials() {
+        return new BasicAWSCredentials(amazonAWSAccessKey, amazonAWSSecretKey);
+    }
+    
+    @Bean
+    public AmazonEC2 amazonEC2() {
+    	AmazonEC2 amazonEC2 = new AmazonEC2Client(amazonAWSCredentials());
+        return amazonEC2;
+    }
+    
+    @Bean
+    public AmazonSQS amazonSQS() {
+    	AmazonSQS amazonEC2 = new AmazonSQSClient(amazonAWSCredentials());
+        return amazonEC2;
+    }
+
+}
